+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Registration module: change output file naming (do not create a registration/ nor registration/transf_matrices/ sub-folders, add _vRG suffix instead of _registered and _transformationMatrix suffixes). Add a new "Output" box in the GUI.
 * Registration module (Alignment tab): search for matching transformation matrices based on unique identifier (part of the basename before the first "_") and warn if multiple matches are found.
 * Event filter module: change output file naming (do not create a event_filter/ sub-folder, add _vEF suffix instead of _mask, _graph and _dictionary suffixes).
+* Registration module: log to file.
 
 ### Removed